refactor(ecau): move maxurl type declarations closer to usage
There's no point to making these declarations available in all modules
when only one module is using them. Also cleans up the declarations
and properly specifies the docs.
